Add `--language` filter option
Closes #11.
